GW-47: Updated partners
