+++ src/pages/partners/Partners.js
@@ -2,9 +2,9 @@ import React from "react";
 import { MainContainer } from "../../shared/layout/MainContainer";
 import { Box } from "@mui/material";
 import { Footer } from "../../shared/components/footer/Footer";
-import SponsorDivider from "./SponsorDivider";
-import SponsorRegistration from "./SponsorRegistration";
-import SponsorTitle from "./SponsorTitle";
+import SponsorDivider from "./components/SponsorDivider";
+import SponsorRegistration from "./components/SponsorRegistration";
+import SponsorTitle from "./components/SponsorTitle";
 import SponsorList from "./components/SponsorList";
 import { LandingImage } from "pages/home/components/LandingImage";
