+++ src/pages/partners/components/SponsorDivider.js
