+++ src/pages/partners/components/SponsorIndex.js
