+++ src/pages/partners/components/SponsorRegistration.js
