+++ src/pages/partners/components/index.js
@@ -1,6 +1,6 @@
 import SponsorContainer from "./SponsorContainer";
 import SponsorDivider from "./SponsorDivider";
-import SponsorList from "./components/SponsorList";
+import SponsorList from "./SponsorList";
 import SponsorRegistration from "./SponsorRegistration"
 import SPONSORS from "./constants";
